Makefile: allow installing without -D
b40a6cb5 (fix parallel install, 2019-02-08), adds -D to the install
calls unconditionally, but that flags is not available in macOS (or
other BSD).

Fix the problem instead by creating the directories as a dependency
and therefore avoiding the need to use -D as well as the multiple
racy calls that will be required to recreate all parent directories
by all install targets.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>